requirements.txt
streamlit
openai
python-dotenv
pillow
numpy
mediapipe==0.10.14
